Routing - Step 3
Here we create a Feed controller. The best way to think about a
Controller is that it is like the highest level component for a route.

We add an `{{outlet}}` to the application route because that's how we
say where to render child routes. The application is the highest level
route so our `feed` route will be rendered inside the `{{outlet}}` of
the application route.

We need to update our link in the header because we're overriding the
default `index` route when we set the `feed` route's path to `'/'`.

Support for converting JavaScript objects to inline styles isn't build
in but there are some easy ways to do that. I'm just going to ignore
that for now because I don't feel like it's important at the moment and
we can revisit that later :-p