Data - Step 1
Here we create the `feed` route. In ember.js, the route is usually
responsible for fetching the data that is needed. Whatever is returned
by the `model()` function is set as the `model` propety of the
controller, which is then available in the `feed.hbs` template as
`model`.

We can then iterate over the `tweets` key of the `model` property in the
template to display the tweets.