Fetching - Step 1-4
Here we create our own Adapter that extend the default `JSONAPIAdapter`
and sets the `host` to point to the local lore tutorial api server.

Then we create the `tweet` model, defining the `text` and `createdAt`
attributes and their types.

We also replace the hard coded data in our `feed` route with an API call
using the ember data store to find all the tweets.

Since the lore-tutorial-api doesn't match the default JSON API format,
we create a custom serializer in order to transform the response from the
server into the expected JSON API format.

Finally we update some templates in order to access data directory off
of the tweet models.