Fetching - Step 5
Here we create the `user` model and declare its attribute schema.

Additionally, we edit the `tweet` model in order to define its
relationship to users. We also update the `tweet` serializer to include
the user relationship data. This allows us to fetch the user for each
tweet through the relationship.

We also create a `user` serializer and override the
`normalizeFindRecordResponse` function in order to transform the data
into the JSON API format expected by ember data.

Finally, we update our templates so that they use the relationship
between tweets and users in order to display the user information for
each tweet.