Fetching - Step 6
Here we modify out `feed` route's model hook to return an `RSVP.hash`,
which is a promise that resolves once all values of its keys (which can
also be promises) resolves. This is really useful because the `model`
hook of a Route will delay rendering until its value is resolved.

Basically, your route's template won't render until all promises
returned by the `model`,`beforeModel`, and `afterModel` hooks have been
resolved.

While any of these promises are being resolved, we'll be in the loading
route substate. Basically, while those promises are being resolved
(while we're fetching our data) we'll enter the loading route. The
loading route's template will be rendered, which is the one we created
here. Once those promises resolve, we'll exit the `loading` route and
enter the `feed` route.